[CDF-20650] - Support injecting custom environmental variables (#272)
* refactor; Moved out fixtures for reuse and approval client into its own module

* refactor: Moved out mocking if environments yaml file

* fix: Empty build directory

* tests: Added failing test for injecting custom variable"

* tests: Extended mocking of CDFToolConfig

* fix; Support custom environment variables

* build: changelog

* tests: regen test data
